+++ c/src/dynamixel_sdk/packet_handler.c
@@ -64,17 +64,8 @@ const char *getTxRxResult(int protocol_version, int result)
     return getTxRxResult2(result);
   }
 }
-void printTxRxResult(int protocol_version, int result)
-{
-  if (protocol_version == 1)
-  {
-    printTxRxResult1(result);
-  }
-  else
-  {
-    printTxRxResult2(result);
-  }
-}
+
+
 const char *getRxPacketError(int protocol_version, uint8_t error)
 {
   if (protocol_version == 1)
@@ -86,17 +77,7 @@ const char *getRxPacketError(int protocol_version, uint8_t error)
     return getRxPacketError2(error);
   }
 }
-void printRxPacketError(int protocol_version, uint8_t error)
-{
-  if (protocol_version == 1)
-  {
-    printRxPacketError1(error);
-  }
-  else
-  {
-    printRxPacketError2(error);
-  }
-}
+
 
 int getLastTxRxResult(int port_num, int protocol_version)
 {